feat: add exemptions search and delete exemption

+++ backend/src/main/java/org/brsm_system_server/repository/ExemptionRepository.java
@@ -3,5 +3,5 @@ package org.brsm_system_server.repository;
 import org.brsm_system_server.entity.Exemption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExemptionRepository extends JpaRepository<Exemption, Integer> {
+public interface ExemptionRepository extends JpaRepository<Exemption, Long> {
 }
